+++ packages/neuron-wallet/src/controllers/app/index.ts
@@ -166,14 +166,6 @@ export default class AppController {
     new UpdateController().checkUpdates(menuItem)
   }
 
-  public static openWebsite() {
-    AppController.openExternal(URL.Website)
-  }
-
-  public static openRepository() {
-    AppController.openExternal(URL.Repository)
-  }
-
   public static showPreference() {
     AppController.navTo(URL.Preference)
   }
@@ -182,8 +174,12 @@ export default class AppController {
     AppController.navTo(URL.CreateWallet)
   }
 
-  public static importWallet() {
-    AppController.navTo(URL.ImportWallet)
+  public static importWallet(type: 'mnemonic' | 'keystore') {
+    if (type === 'mnemonic') {
+      AppController.navTo(URL.ImportMnemonic)
+    } else if (type === 'keystore') {
+      AppController.navTo(URL.ImportKeystore)
+    }
   }
 
   public static async showTransactionDetails(hash: string) {