+++ packages/neuron-wallet/src/locales/zh.ts
@@ -20,6 +20,8 @@ export default {
         backup: '备份当前钱包',
         delete: '删除当前钱包',
         'change-password': '修改密码',
+        'import-mnemonic': '导入助记词',
+        'import-keystore': '导入 Keystore 文件',
       },
       edit: {
         label: '编辑',
@@ -40,7 +42,8 @@ export default {
       },
       help: {
         label: '帮助',
-        sourceCode: '源代码',
+        'source-code': '源代码',
+        'report-issue': '报告问题',
         settings: '设置',
       },
       develop: {
@@ -89,6 +92,8 @@ export default {
       'capacity-not-enough': '余额不足',
       'capacity-too-small': '金额小于最低金额',
       'should-be-type-of': '{{field}} 应该为 {{type}} 类型',
+      'invalid-keystore': 'Keystore 格式不正确',
+      'invalid-json': 'JSON 文件格式不正确',
     },
     contextMenu: {
       select: '选择',
@@ -136,7 +141,7 @@ export default {
     },
     updater: {
       'update-not-available': '没有可供升级的新版本。',
-      'updates-found-do-you-want-to-update': '有可供升级的新版本。马上进行升级吗？',
+      'updates-found-do-you-want-to-update': '有可供升级的新版本({{version}})。现在进行升级吗？',
       'update-now': '马上升级',
       'updates-downloaded-about-to-quit-and-install': '下载完成。Neuron 将退出并安装新版本...',
     },